Made Changes

+++ PaymentService/src/main/java/org/padmini/railway/dao/UserPaymentRepository.java
@@ -3,7 +3,7 @@ package org.padmini.railway.dao;
 import org.padmini.railway.entity.PaymentDetails;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface UserPaymentRepository extends MongoRepository<PaymentDetails, Integer>
+public interface UserPaymentRepository extends MongoRepository<PaymentDetails, Long>
 {
 
 }
